Fix hotspot volunteer name update operation
Updates for the name of a volunteer associated with the hotspot are not displayed immediately. Unfortunately, attempting to use Repo.preload to fix this does not work as expected. The update has to be forced by inserting the new volunteer values into the returned entity, to avoid having a stale state returned to the front-end.